refactor: change route-config to route-constant
