+++ apps/web/src/services/administration/components/RoleCreateEditHandbook.vue
@@ -67,7 +67,7 @@
 import { PLink, PDivider } from '@spaceone/design-system';
 import { ACTION_ICON } from '@spaceone/design-system/src/inputs/link/type';
 
-import { ADMINISTRATION_ROUTE } from '@/services/administration/routes/route-config';
+import { ADMINISTRATION_ROUTE } from '@/services/administration/routes/route-constant';
 
 export default {
     name: 'RoleCreateEditHandbook',
