+++ apps/web/src/services/administration/components/RoleManagementTab.vue
@@ -70,7 +70,7 @@ import { i18n } from '@/translations';
 
 import RoleManagementTabDetail from '@/services/administration/components/RoleManagementTabDetail.vue';
 import { ROLE_TYPE_BADGE_OPTION } from '@/services/administration/constants/role-constant';
-import { ADMINISTRATION_ROUTE } from '@/services/administration/routes/route-config';
+import { ADMINISTRATION_ROUTE } from '@/services/administration/routes/route-constant';
 import { useRolePageStore } from '@/services/administration/store/role-page-store';
