+++ apps/web/src/services/administration/pages/PolicyCreatePage.vue
@@ -75,7 +75,6 @@ import {
 
 import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 
-import { POLICY_TYPE } from '@/api-schema/identity/policy/model';
 import { SpaceRouter } from '@/router';
 import { i18n } from '@/translations';
 
@@ -84,7 +83,9 @@ import { showSuccessMessage } from '@/lib/helper/notice-alert-helper';
 import ErrorHandler from '@/common/composables/error/errorHandler';
 import { useFormValidator } from '@/common/composables/form-validator';
 
-import { ADMINISTRATION_ROUTE } from '@/services/administration/routes/route-config';
+import { POLICY_TYPE } from '@/services/administration/constants/policy-constant';
+import { ADMINISTRATION_ROUTE } from '@/services/administration/routes/route-constant';
+
 
 export default {
     name: 'PolicyCreatePage',
